+++ CHANGELOG.md
@@ -82,6 +82,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2293 Migrate Average Glucose from stat.js to yourloops
 - YLP-2300 Remove the remote monitoring invitation flow
 - YLP-2301 Split ITs jobs by roles
+- YLP-2307 Incomplete alert when accepting a direct share invitation
 - YLP-2311 Reorganize ITs
 - YLP-2367 Patient list refacto